cli/requirements.txt
# CLI Dependencies
google-cloud-storage==2.20.0
requests==2.32.3
openai==1.59.4

# Import backend dependencies
-r ../backend/requirements.txt
